# swagger.yaml
openapi: 3.0.3
info:
  title: MeetFlow API
  version: 1.0.0
  description: |
    ## MeetFlow Backend REST API

    This API powers the MeetFlow SaaS platform. It is divided into four
    top-level sections:

    | Section          | Base Path          | Description                                    |
    |------------------|--------------------|------------------------------------------------|
    | **Common**       | `/api/v1/auth` & `/api/v1/account`          | Authentication & account management            |
    | **Master**       | `/api/v1/master`   | MeetFlow internal admin panel (master context) |
    | **Organization** | `/api/v1/organization` | Organisation-level management (coming soon) |
    | **Workspace**    | `/api/v1/workspace`    | Workspace & channel management (coming soon)|

    ### Authentication
    All **Master**, **Organization**, and **Workspace** routes require a valid
    JWT `accessToken` in the `Authorization` header:
    ```
    Authorization: Bearer <accessToken>
    ```
    Use the **Login** endpoint to obtain tokens.

    ### Standard Response Envelope
    Every response (success or error) is wrapped in:
    ```json
    {
      "status": {
        "response_code": 200,
        "response_message": "Success!"
      },
      "message": "Human-readable message",
      "data": { ... }
    }
    ```
    Error responses replace `data` with an `error` object containing `details`.

  contact:
    name: MeetFlow Engineering
  license:
    name: ISC
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT access token returned by POST /auth/login.
        Valid for 1 day (or 30 days when rememberMe is true).
  parameters:
    PathId:
      name: id
      in: path
      required: true
      description: Numeric record ID (BIGINT primary key)
      schema:
        type: integer
        format: int64
        example: 1
    QuerySearch:
      name: search
      in: query
      required: false
      description: Full-text search / keyword filter
      schema:
        type: string
        example: admin
    QueryPage:
      name: page
      in: query
      required: false
      description: Page number for pagination (1-based)
      schema:
        type: integer
        default: 1
        minimum: 1
    QueryLimit:
      name: limit
      in: query
      required: false
      description: Number of records per page
      schema:
        type: integer
        default: 10
        minimum: 1
        maximum: 100
    QueryIsActive:
      name: is_active
      in: query
      required: false
      description: Filter by active / inactive status
      schema:
        type: boolean
  schemas:
...
    StatusBlock:
      type: object
      properties:
        response_code:
          type: integer
          example: 200
        response_message:
          type: string
          example: "Success!"
    SuccessEnvelope:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/StatusBlock"
        message:
          type: string
          example: "Operation successful"
        data:
          description: Payload — varies per endpoint
          nullable: true
    ErrorEnvelope:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/StatusBlock"
        message:
          type: string
          example: "Something went wrong!"
        error:
          type: object
          properties:
            details:
              type: string
              example: "Invalid credentials."
    DeletedResponse:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - type: object
          properties:
            data:
              nullable: true
              example: null

    # ── Auth ──────────────────────────────────────────────────────────────────
...
    Role:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          maxLength: 70
          example: master_admin
        description:
          type: string
          nullable: true
          example: "Full access to master panel"
        is_system:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateRoleRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          maxLength: 70
          example: master_admin
        description:
          type: string
          example: "Full access to master panel"
        is_system:
          type: boolean
          default: true
    UpdateRoleRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 70
        description:
          type: string
        is_system:
          type: boolean

    # ── Master User ───────────────────────────────────────────────────────────
...
    BulkCreateRoleItem:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          maxLength: 70
          example: custom_admin
        description:
          type: string
          example: "Custom admin privileged role"
        is_system:
          type: boolean
          default: true
    BulkCreateRoleRequest:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          minItems: 1
          maxItems: 100
          items:
            $ref: "#/components/schemas/BulkCreateRoleItem"
    BulkUpdateRoleItem:
      type: object
      required:
        - id
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          maxLength: 70
        description:
          type: string
        is_system:
          type: boolean
    BulkUpdateRoleRequest:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          minItems: 1
          maxItems: 100
          items:
            $ref: "#/components/schemas/BulkUpdateRoleItem"
    BulkDeleteRoleRequest:
      type: object
      required:
        - ids
      properties:
        ids:
          type: array
          minItems: 1
          maxItems: 100
          items:
            type: integer
            format: int64
          example: [1, 2, 3]

    # ── Bulk Users ────────────────────────────────────────────────────────────
...
  responses:
    UnauthorizedError:
      description: Missing or invalid access token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            status:
              response_code: 401
              response_message: "Something went wrong!"
            message: "Access token is required."
    ForbiddenError:
      description: Authenticated but insufficient permissions (wrong context)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            status:
              response_code: 403
              response_message: "Something went wrong!"
            message: "This route is only accessible to master users."
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    BadRequestError:
      description: Validation error or bad input
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
paths:
...
  /master/roles/get-all-roles:
    get:
      tags: [Master / Roles]
      summary: Get all roles
      operationId: getAllRoles
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/QuerySearch"
        - $ref: "#/components/parameters/QueryPage"
        - $ref: "#/components/parameters/QueryLimit"
        - $ref: "#/components/parameters/QueryIsActive"
      responses:
        "200":
          description: Roles retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Role"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
  /master/roles/get-role-by-id/{id}:
    get:
      tags: [Master / Roles]
      summary: Get role by ID
      operationId: getRoleById
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
      responses:
        "200":
          description: Role retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Role"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
  /master/roles/create-role:
    post:
      tags: [Master / Roles]
      summary: Create a role
      operationId: createRole
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRoleRequest"
            example:
              name: master_admin
              description: "Full access to master panel"
              is_system: true
      responses:
        "201":
          description: Role created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Role"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
  /master/roles/update-role/{id}:
    patch:
      tags: [Master / Roles]
      summary: Update a role
      operationId: updateRole
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRoleRequest"
      responses:
        "200":
          description: Role updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Role"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
  /master/roles/delete-role/{id}:
    delete:
      tags: [Master / Roles]
      summary: Soft delete a role
      operationId: deleteRole
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
      responses:
        "200":
          description: Role deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeletedResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"

  # ─── Bulk Role Operations ──────────────────────────────────────────────────
  /master/roles/bulk-create:
    post:
      tags: [Master / Roles]
      summary: Bulk create roles
      description: |
        Create multiple roles in a single request.
        **Limits:** Maximum 100 items. Each name must be unique.
      operationId: bulkCreateRoles
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkCreateRoleRequest"
      responses:
        "200":
          description: Bulk create completed
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          successful:
                            type: array
                            items:
                              $ref: "#/components/schemas/Role"
                          failed:
                            type: array
                          summary:
                            type: object
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
  /master/roles/bulk-update:
    patch:
      tags: [Master / Roles]
      summary: Bulk update roles
      description: |
        Update multiple roles in a single request.
        Each item must include an `id`. Other fields are optional.
        **Limits:** Maximum 100 items.
      operationId: bulkUpdateRoles
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkUpdateRoleRequest"
      responses:
        "200":
          description: Bulk update completed
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        type: object
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
  /master/roles/bulk-delete:
    delete:
      tags: [Master / Roles]
      summary: Bulk delete roles
      description: |
        Soft delete multiple roles by ID.
        **Limits:** Maximum 100 IDs.
      operationId: bulkDeleteRoles
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkDeleteRoleRequest"
      responses:
        "200":
          description: Bulk delete completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"

  # ═══════════════════════════════════════════════════════════════════════════
  # MASTER — Users
  # ═══════════════════════════════════════════════════════════════════════════
...